feat(root): update-package-publishing
update package publishing.

we updated the way we bundle and package the widget so that we can
support esm and cjs builds. we generate types through svelte-kit package
which then gets bundled in a seperate rollup config. We then
post-process the types so that we can remove namespaces that are not
wanted due to the nature of generating types (svelte).