fix(root): update publishing to add readme and changelog
a release should include a readme.md and a changelog.md file. we are generating that with semantic
release so its important that its generated in the package/* directory so that its automatically
picked up by the npm publish. then we can include such artifacts in the github release assets