fix: accessible-fields-when-data-is-unchanged
fields that require focus when no data has changed can be missed
because svelte wont re run components or data where things are not
changed.
we can force this by using the key component wrapper.
this is a way to override the default behavior when a value is changed,
so the entire component triggers